add modules for navigator, set up router file in config folder of components, refactor bracket components, rename files for better naming convention and navigating
