+++ package.json
@@ -15,7 +15,9 @@
     "morgan": "^1.8.2",
     "react": "16.0.0-alpha.12",
     "react-native": "0.47.1",
+    "react-native-elements": "^0.15.0",
     "react-native-keychain": "^2.0.0-rc",
+    "react-navigation": "^1.0.0-beta.11",
     "react-redux": "^5.0.5",
     "react-tournament-bracket": "^0.1.0",
     "redux": "^3.7.2",
